refactor: create a new application layer
